feat: category filter 클릭 시 필터가 동작하도록 설정
